+++ src/generator/none.cpp
@@ -1,5 +1,5 @@
 #include "none.hpp"
 
 NoGenerator::NoGenerator() {}
-float NoGenerator::getSample(float phase, float t) { return 0; }
+double NoGenerator::getSample(double phase, double t) { return 0; }
 Generator* NoGenerator::copy() {return new NoGenerator();}
