+++ src/generator/triangle.hpp
@@ -4,16 +4,16 @@
 #include "generator.hpp"
 
 class TriangleGenerator :public Generator {
-    Interpolated<float> m_peak;
+    Interpolated<double> m_peak;
     
 public:
-    TriangleGenerator(Interpolated<float> amplitude = 1.0f, 
-                      Interpolated<float> phasemul = 1.0f, 
-                      Interpolated<float> offset = 0.0f, 
-                      Interpolated<float> peak = 0.5f);
+    TriangleGenerator(Interpolated<double> amplitude = 1.0f, 
+                      Interpolated<double> phasemul = 1.0f, 
+                      Interpolated<double> offset = 0.0f, 
+                      Interpolated<double> peak = 0.5f);
     TriangleGenerator(std::istream& stream);
 
-    float getSample(float phase, float t);
+    double getSample(double phase, double t);
     Generator* copy();
 };
